+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 * basic main menu
 * main function to start creating character
 * user can give a name to character
+* add more character and race attributes: size, speed, vision